placeOrders.py
from alpaca.trading.requests import MarketOrderRequest
from alpaca.trading.enums import OrderSide, TimeInForce

# Define the market order
market_order = MarketOrderRequest(
    symbol="AAPL",
    qty=1,
    side=OrderSide.BUY,
    time_in_force=TimeInForce.GTC
)

# Submit the order
order = trading_client.submit_order(order_data=market_order)
print(order)
