requirements.txt
alpaca-trade-api==3.0.0
scikit-learn==1.3.0
pandas==2.1.1
ta==0.10.2
python-dotenv==1.0.0
